Move randomUUID from a  user defined function to a builtin function ()
randomUUID() function is no longer case sensitive
