+++ community/cypher/front-end/expressions/src/main/scala/org/neo4j/cypher/internal/expressions/functions/Function.scala
@@ -67,6 +67,7 @@ object Function {
     Keys,
     Radians,
     Rand,
+    RandomUUID,
     Range,
     Reduce,
     Relationships,
