+++ community/cypher/interpreted-runtime/src/main/scala/org/neo4j/cypher/internal/runtime/interpreted/commands/expressions/Expression.scala
@@ -55,6 +55,7 @@ abstract class Expression extends AstNode[Expression] {
 
   val isDeterministic: Boolean = ! exists {
     case RandFunction() => true
+    case RandomUUIDFunction() => true
     case _              => false
   }
 }
